Make contact matrix from Mistry et al. data

Make contact matrix from Mistry et al. data

Usage

mk_contact_pars(age.group.lower, setting.weight)

Arguments

age.group.lower

Numeric vector. Lower bounds for desired age groups

setting.weight

Named numeric vector of the average contact rate by setting to reflect closures.

Value

A list with:

p.mat: the row-normalized matrix of contact probabilities

c.hat: the vector of row sums from the contact rate matrix, giving the average contact rate per age group; used to set up the transmission vector to maintain the contact balance condition